Update Training.py
Removing the unnecessary function for regularization term in the loss regarding the RBF. This term was one of the initial ideas to ensure a balanced RBF response across centers for each data point, preventing any single center from overly dominating the output. This design was supposed to ensure that all data points achieve a high RBF score, enhancing anomaly detection capabilities. However it is not considered in final work.